# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "NTASK-API",
        "description": "NTASK MANAGER NODEJS API",
        "version": 1.0
    },
    "schemes": ["http", "https"],
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication endpoint"
        },
        {
            "name": "User",
            "description": "User endpoint"
        },
        {
            "tasks": "Task",
            "description": "Task endpoint"
        }
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "JWT Bearer token"
        }
    },
    "definitions": {
        "affectedCount": {
            "type": "array",
            "items": {
                "type": "number"
            }
        },
        "forgotPassword": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "login": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            },
            "required": ["email", "password"]
        },
        "createUser": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            },
            "required": ["name", "email", "password"]
        },
        "responseUser": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            },
            "required": ["name", "email", "password"]
        },
        "responseTask": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "userId": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "bool"
                },
                "tag": {
                    "type": "string"
                }
            },
            "required": ["name", "email", "password"]
        },
        "createTask": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                }
            },
            "required": ["title", "tag"]
        },
        "updateTask": {
            "type": "object",
            "properties": {
                "title": "title",
                "tag": "string",
                "completed": "bool"
            },
            "required": ["title", "tag", "completed"]
        }
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": ["Authentication"],
                "summary": "Authenticate user",
                "description": "Authenticate user and return a json-web-token",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User login information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentication Successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "500": {
                        "description": "Internal Server error"
                    }
                }
            }
        },
        "/auth/forgotPassword": {
            "post": {
                "tags": ["Authentication"],
                "summary": "Send reset a token to user",
                "description": "Validate the email and send a json-web-token to user email",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "email",
                        "description": "User email information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forgotPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email sent Successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server error"
                    }
                }
            }
        },
        "/auth/resetPassword/:token": {
            "post": {
                "tags": ["Authentication"],
                "summary": "Reset user password",
                "description": "Validate the json-web-token and reset the user password",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "password",
                        "description": "New user password information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forgotPassword"
                        }
                    },
                    {
                        "in": "path",
                        "name": "token",
                        "description": "user reset token sent to email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email sent Successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server error"
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "tags": ["User"],
                "summary": "Create a User",
                "description": "Create a user and return a UserDTO",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User create information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Register Successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "$ref": "#/definitions/responseUser"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },

                    "500": {
                        "description": "Internal Server error"
                    }
                }
            }
        },
        "/user/getById": {
            "get": {
                "tags": ["User"],
                "summary": "Get a User",
                "description": "Find a user by id and return a UserDTO",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search Successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "$ref": "#/definitions/responseUser"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server error"
                    }
                }
            }
        },
        "/user/delete": {
            "delete": {
                "tags": ["User"],
                "summary": "Delete a User",
                "description": "Find a user by id delete and return a the result",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successful"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server error"
                    }
                }
            }
        },
        "/task/create": {
            "post": {
                "tags": ["Task"],
                "summary": "Create a Task",
                "description": "Create a task and return a UserDTO",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Task create information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/responseTask"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "$ref": "#/definitions/responseUser"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server error"
                    }
                }
            }
        },
        "/task/getAll": {
            "get": {
                "tags": ["Task"],
                "summary": "Get all Task",
                "description": "Find all tasks and return",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search Successful",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responseTask"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server error"
                    }
                }
            }
        },
        "/task/getById/{id}": {
            "get": {
                "tags": ["Task"],
                "summary": "Get a Task",
                "description": "Find a task by id and return",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "task id to retrieve",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search Successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "$ref": "#/definitions/responseTask"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server error"
                    }
                }
            }
        },

        "/task/update/{id}": {
            "update": {
                "tags": ["Task"],
                "summary": "Get a Task",
                "description": "update a task by id and return",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "task id to update",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated Successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "$ref": "#/definitions/responseTask"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server error"
                    }
                }
            }
        },
        "/task/delete": {
            "delete": {
                "tags": ["Task"],
                "summary": "Delete a Task",
                "description": "Find a task by id delete and return the result",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successful"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server error"
                    }
                }
            }
        }
    }
}